Implement :xc[lip] and :nerdlog commands
This is a simple mechanism to share log queries with each other:
e.g. when we're using web-based tools, we usually can just copy the URL
and paste it to our coworkers; the equivalent in nerdlog is now `:xclip`
command, which copies to clipboard a shell-ready command like this:

```
nerdlog --hosts 'mynode-*' --time -5h --query '/ping/'
```

And then others can either just run it in the terminal, OR they can
paste it straight in the already opened nerdlog instance, because
`:nerdlog` is also a supported command, so just hit the `:` and paste it
there:

```
:nerdlog --hosts 'mynode-*' --time -5h --query '/ping/'
```

And you'll get the same result.