Gzip data on the server, to reduce amount of traffic
This actually speeds up queries on my machine: what used to be about 3s,
now takes about 2s. I don't yet have exact numbers as to what the amount
of downloaded data is before or after this change, but just looking at a
single node, the difference on it is like 7KB vs 58KB, so it's a pretty
significant drop in traffic.