In nerdlog_query.sh, format time in local timezone
The problem is that while awk's strftime supports formatting time in
UTC, mktime doesn't (so it always parses time in local timezone), so we
can't make it work reliably in awk.

So the idea is that on the nodes, we'll always be using local time, and
then the nerdlog app will just initially query the timezone from the
node, and will be converting all timestamps for this particular node
to/from that timezone (that part is still TODO).