Improve error handling and agent script output parsing
First, we now wait for the command_done marker to be printed on both
stdout and stderr; otherwise there was a race: we could have read that
marker from stdout, and conclude the command before reading all the
lines from stderr.

Second, we now collect all the lines starting with "error:" at both
stdout and stderr univerally for all commands.

Third, we also collect unhandled stdout and stderr lines, and if there
are no exlicit errors ("error:" lines), we then use the raw stdout and
stderr in the error message.

Fourth, we now make use of the exit code returned from the script, even
in the query case, where we pipe the agent output to gzip.