Begin refactoring Go message parsing
The general idea is to have a LogMsg from the very beginning, and then
improve it in a few parsing stages: first parse timestamp, then envelope
(e.g. app name etc from the syslog message), then the app-level payload.

Not everything is done yet, but we have the timestamp parsing and custom
parsing in this pipeline.