Handle Alt+Left and Alt+Right for navigating history
We already had `:back` and `:fwd` to navigate browser-like history, but
it wasn't very ergonomic, so in attempt to address that, this commit
adds a handler for Alt+Left and Alt+Right, just like the typical
shortcuts in a browser.

Why not `Ctrl+K` and `Ctrl+J` as on the query edit form: because on that
form, Ctrl+K and Ctrl+J behave like a command-line history rather than a
browser-like history. Consider:

If we have an query history like this:

- `/foo/`
- `/bar/`
- `/baz/` (current query)

If we now go back 2 times (to `/foo/`), and after that execute a new
query `/xyz/`, then the browser-like history will look like this:

- `/foo/`
- `/xyz/` (current query)

So as you see, since we went back to `/foo/` and then executed a new
query after that, then both `/bar/` and `/baz/` are dropped.

While the commandline-like history will rather look like this:

- `/foo/`
- `/bar/`
- `/baz/`
- `/bar/`
- `/foo/`
- `/xyz/` (current query)

As you see, it's just a linear history of queries as they happened in
time.

So to avoid adding confusion about what `Ctrl+K` and `Ctrl+J` mean, it's
better to use different shortcuts for the browser-like history.