Rename cmd/nerdlog-tui to just cmd/nerdlog
When installing binaries using `go install .....`, the resulting binary
has the same name as the directory, so it used to be `nerdlog-tui`,
which I think is a bit of a mouthful.

So since as of today TUI is the only flavor of nerdlog, I think it's a
better tradeoff to just have it named `nerdlog`.