Allow reading log files with sudo
As reported in https://github.com/dimonomid/nerdlog/issues/9, on some
systems the log files are owned by `root:root`, so the only way to read
these files is to either login as root, or to use `sudo` (or to use
`journalctl`, but it's not yet supported by nerdlog).

So as an (insecure) workaround, this commit adds support for using
`sudo` when reading log files:

First, make sure that the nerdlog agent script can run with sudo without
password, by adding something like this to the `/etc/sudoers`:

```
myuser  ALL=(ALL) NOPASSWD: /tmp/nerdlog_agent*
```

And then add `options: {"sudo": true}` to `/etc/logstreams.yaml`
for the corresponding logstream, like this:

```yaml
log_streams:
  myhost-01:
    # ... Potentially any other configuration for the logstream
    options: {"sudo": true}
```

Added docs for that to
https://dmitryfrank.com/projects/nerdlog/article#reading_log_files_with_sudo

As a side change here, we now make the agent script executable, and run
it directly without `bash`, just to make it possible to configure
`/tmp/nerdlog_agent*` in the sudoers file (although it's still very
insecure, since the user can just put anything in that file).

We can potentially add a way to use some previously-provisioned agent
script owned by root, instead of uploading it on every connection: this
way, we can that script in the sudoers file, and therefore it'll be more
secure, but it's also not ideal since the protocol will likely change in
further versions of nerdlog, which means maintenance burden. So not sure
if it's worth, and not bothering for now.