Refactor nerdlog_agent.sh as preparation for journalctl support
Previously, the logic was: first make sure that the index is present and
valid, and then run the main awk script to generate the actual data for
the client.

However, for journalctl, the indexing part will not be needed (not in
the existing form anyway), so to make it easier to run the main awk
script without the indexing, running that main awk script is now
factored out as a bash function which takes a few arguments as env vars.