Fix focus when non-last modal is removed
It was a long overdue bug:

- We make a query, so the query-in-progress dialog appears
- Something else happens and we get another modal on top
- The query is finished, so the query-in-progress dialog disappears

In this case, the focus was set to the main window, and it's not even
usable because it's not the current "page" as tview calls it, so the UI
was stuck.

This commit fixes it so that when non-last modal window is removed, we
don't change the focus, and also update the focus history so that when
the next modal is also removed at some point, it'll give back the focus
to the right primitive (not to the query dialog which was just removed).