fix(UI): Fix timezone label in query edit form
It was misleadingly saying that the timestamps should be in UTC, but
it's not true: it was a remnant from the old times when all times in
Nerdlog UI were indeed in UTC, but now it's configurable with
`:set timezone=`.

So this commit removes that, and adds a separate label at the right of
the input with the actual offset being used.