feat: Implement proper support for localhost
Previously, localhost logstreams were still going through ssh, which
meant that to read local logs one would need to have ssh daemon running
locally, which is uncommon for developer machines.

So this commit fixes it by implementing `ShellTransportLocal` and using
it if the hostname is literally `localhost` (in which case, port and
user are ignored).