feat: Don't fail if there is a Match directive in ssh config
Related issue: https://github.com/dimonomid/nerdlog/issues/12

We're far from addressing this issue completely, since it'd be a lot of
effort; but at least to make it less of a problem, this commit works it
around in most cases by ignoring the `Match` directive, since it's
unlikely to have any real effect for the Nerdlog's needs.

If the user actually has this `Match` directive, Nerdlog prints a
warning, which can also be suppressed by setting the
`NERDLOG_NO_WARN_SSH_MATCH` env var.