feat: Handle decreased timestamps gracefully
Previously, if a log file contained decreased timestamps, Nerdlog was
detecting that and refusing to work.

However, it does happen sometimes that the timestamps are decreasing, so
this commit implements more graceful handling of these: a decreased
timestamp is treated as if the timestamp did not change since the
previous line.

There are still some corner cases, e.g. if we filter out some lines
*before* the ones with the decreased timestamp, then the effective
timestamp of the problematic lines might also change (it'll always be
the same as the timestamp of the previous line that we're taking into
account). There are ways to address this too, but it would reduce the
performance (and also add more complexity), which I think is not worth
for this corner case.