test: Add tests for journalctl code in the agent script
No new bugs were discovered (except for one very minor non-issue that
`p:stage:4:done` was not printed in the end; didn't affect anything in
practice).

There is a known issue that in a multi-line message, the lines end up
reversed. Fixing it will be a subject of some other commit.