fix: Fix pagination of journalctl-powered logstreams
It was possible that some messages get lost in between of the already
loaded data, and the newly fetched data.

Pagination with journalctl is kind of a pain; even though there is a
built-in mechanism for it with the `--cursor` and related flags, it only
works if we use solely journalctl to apply patterns (using e.g.
`--grep`) and limit number of messages (using e.g. `--lines`). Then
yeah, we can also make use of `--show-cursor`, and then use `--cursor`
when fetching the next page.

However, Nerdlog applies patterns and limits the number of messages
externally, piping the output to awk. It would be possible if journalctl
could show the cursor for every single message, but sadly it's only
possible to show it for the latest message in the output, so there's no
reliable way to filter some lines out, and get the exact number of lines
with the cursor of the latest one.

Another possible idea is to actually use `--grep` when working with
journalctl, but it also falls apart because the patterns won't be
compatible between journalctl-powered and file-powered logstreams; and
also tbh awk patterns are a lot more convenient due to the support of
multiple patterns with boolean operators.

So long story short, we have to do some hacks with the timestamps to do
the pagination, and I originally overestimated the precision at which
journalctl indexes logs: even though it outputs logs with the
microsecond precision, it's not possible to reliably query logs starting
from this exact microsecond timestamp. Apparently the time bounds need
to be specified with the second precision: then it works correctly, at
least in my experiments so far (journalctl docs don't explicitly mention
the accuracy of the internal index, but at least the examples do use
seconds, so it's reasonable to assume that we can actually rely on the
second precision).

This commit fixes the bug by always specifying the time bounds with the
second precision, and then manually filtering out the messages outside
of the range that we actually need.