feat: Add support for shell init commands in logstream config
The use case right now is only for tests, to set env vars like
`export TZ=UTC`, but might be useful in real life as well.

Usage is as follows:

```
log_streams:
  myhost-01:
    # ... Potentially any other configuration for the logstream
    options:
      shell_init:
        - 'export TZ=UTC'
        - 'export FOO=bar'
        - 'some other command || exit 1'
```

If a command exits with non-zero code (like we do in the example above
with `exit 1`), the whole bootstrapping is failed.