refactor: Allow time to be mocked in core
This will be obviously used in tests.

Also, always pass `CUR_YEAR` and `CUR_MONTH` env vars to agent script.
Outside of tests it's not actually necessary to pass these env vars,
since the agent script will just use the actual time then; but for
simplicity, and to make the tested code closer to the actual code, we
always pass them when calling agent script.