fix: Support ancient bash shipped with MacOS
MacOS is shipped with ancient bash 3.2.x, which doesn't handle `read -r`
in the same way as bash 5.x.x, and there was a bug (in the agent script)
which didn't make a difference in recent bash, but it does break things
on older bash.

So this commit fixes it, and now Nerdlog can read logs from MacOS with
the preinstalled bash.