fix: Forget queued commands when reconnecting
`LStreamClient` is able to queue commands to be executed once the
current command is done, but these queued commands need to be forgotten
when we're reconnecting, and they weren't, so this commit fixes that.

Ideally I also want to add a test for it, but it requires implementing
support for custom ssh session start command (instead of just shell),
which requires more thought, so it's left for later.