fix: Use /bin/sh as the main ssh session shell
Previously nerdlog was using the login shell via ssh (which was causing
problems with shells like `fish`), and locally it was using
`/usr/bin/env bash`, which is a bit insonsistent with the ssh conns.

Note that we can't remove the dependency on `bash` completely yet,
because the agent script still relies on bash-specific features and it's
not easy to change (e.g. I do want to use the `PIPESTATUS` and there's
no real equivalent in POSIX). But at least the main ssh session doesn't
need any bash-specific features, and using `/bin/sh` is the most
reasonable thing there.