feat: Add support for --set command line flag
Which can be used for setting nerdlog options, using the same syntax as
the `:set` command, so e.g. `--set numlines=1000`.

It provides some mitigation for not yet having a persistent config file,
which requires more thought if we want it to be a script and not just a
config file.

It'll be useful even when we have the config file as well though, since
it will take precedence over options in the config file. In particular
it will likely be used in tests.