fix: Fix going from May to Jun in traditional syslog format
There was code to generate better histogram data in a corner case when
the timestamp decreases, but unfortunately it broke a very common case
when the timestamps are going from May to Jun.

So this commit disables the corner-case handling for now (until we have
a better solution), in order to fix the May to Jun transition.