test: Reenable tests for decreased timestamps
I experimented a bit with implementing this support of decreased
timestamps properly, but for that we'd have make awk-side timestamp
parsing a lot heavier, which does slow things down noticeably, so for
now I decided to just give up and accept non-ideal behavior with
decreased timestamps. Slowing things down just to handle this corner
case doesn't seem justified.

It might be implemented later as an optional feature (although also not
sure if it's worth).